core/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
google-genai>=1.0.0
google-auth>=2.23.0
aiohttp>=3.9.0
PyYAML>=6.0
pytest>=7.4.0
httpx>=0.28.0
openpyxl>=3.1.0
reportlab>=4.0.0
redis>=5.0.0
Pillow>=10.0.0
alembic>=1.12.0
bcrypt>=4.2.0
PyJWT>=2.8.0